fix(sonar): clear small Web/Java/TS smells outside S7924
Fix multi-declare, dead conditions, useless assigns, autocomplete, type aliases, dataset/export patterns, video tracks, and a batch of one-off Java/TS maintainability findings.

+++ services/trade-service/src/main/java/com/aicabinet/trade/service/MerchantPortalService.java
@@ -296,7 +296,7 @@ public class MerchantPortalService {
         if (merchantIds != null && !merchantIds.isEmpty()) {
             for (OrderRevenueSplit split : splitRepository.findByMerchantIdInAndCreatedAtAfter(merchantIds, since)) {
                 LocalDate day = split.getCreatedAt().atZone(zone).toLocalDate();
-                incomeByDay.merge(day, (long) split.getMerchantCents(), Long::sum);
+                incomeByDay.merge(day, split.getMerchantCents(), Long::sum);
             }
         }
